third_party/WebKit/LayoutTests/inspector/console/console-uncaught-promise.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/console-test.js"></script>
<script>

function runPromises(source)
{
    Promise.reject(new Error(source + ".err1"))
        .then()
        .then()
        .then(); // Last is unhandled.

    var reject
    var m0 = new Promise(function(res, rej) { reject = rej; });
    var m1 = m0.then(function() {});
    var m2 = m0.then(function() {});
    var m3 = m0.then(function() {});
    var m4 = 0;
    m0.catch(function() {
        m2.catch(function() {
            m1.catch(function() {
                m4 = m3.then(function() {}); // Unhandled.
            });
        });
    });
    reject(new Error(source + ".err2"));
}

function onload()
{
    runPromises("onload");
    runTest();
}

function runPromisesFromInspector()
{
    // setTimeout to cut off VM call frames from the stack trace.
    setTimeout(function timeout() {
        runPromises("inspector")
    }, 0);
}

function test()
{
    InspectorTest.addConsoleViewSniffer(checkConsoleMessages, true);
    WebInspector.console.showPromise().done();

    checkConsoleMessages();

    function checkConsoleMessages()
    {
        var count = InspectorTest.consoleMessagesCount();
        if (count === 2)
            InspectorTest.evaluateInPage("runPromisesFromInspector()");
        else if (count === 4)
            InspectorTest.expandConsoleMessages(dump);
    }

    function dump()
    {
        InspectorTest.dumpConsoleMessages(false, false, InspectorTest.textContentWithLineBreaks);
        InspectorTest.completeTest();
    }
}

</script>
</head>

<body onload="onload()">
<p>
Tests that uncaught promise rejections are logged into console.
</p>

</body>
</html>
